@@ -1,25 +1,44 @@
 {
+  "name": "simple-crud-api",
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev:debug": "nodemon --config nodemon.json --inspect-brk src/server.ts",
+    "start": "ts-node  src/server.ts",
+    "dev": "nodemon  --exec  \"ts-node  src/server.ts\" ",
+
+    "lint": "eslint src",
+    "start:dev": "nodemon",
+    "prod": "tsc && node dist/server.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/spadarynjaALINA/Simple-CRUD-API.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/spadarynjaALINA/Simple-CRUD-API/issues"
+  },
+  "homepage": "https://github.com/spadarynjaALINA/Simple-CRUD-API#readme",
   "devDependencies": {
+    "@types/node": "^18.0.0",
+    "eslint": "^8.18.0",
     "nodemon": "^2.0.16",
+    "ts-node": "^10.8.1",
+    "typescript": "^4.7.4",
     "webpack": "^5.73.0",
     "webpack-cli": "^4.10.0"
   },
-  "scripts": {
-    "start":"node server.js",
-    "dev":"nodemon server.js",
-    "build": "webpack",
 
-    "lint": "eslint src"
-  },
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.28.0",
-    "@typescript-eslint/parser": "^5.28.0",
-    "eslint": "^8.17.0",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "prettier": "^2.7.0",
-    "typescript": "^4.7.3",
-    "webpack-dev-server": "^4.9.2"
+    "@types/uuid": "^8.3.4",
+    "dotenv": "^16.0.1",
+    "esm": "^3.2.25",
+    "prettier": "^2.7.1",
+    "uuid": "^8.3.2"
   }
 }
